fix internal term links; add inline code spans
now <<::term>> from anywhere goes to the term in the right page.
and titles of internal terms links are fixed (eg <<::check_out>> would
now result in an <a ...>check out</a>, not an <a ...>check_out</a>).

``code`` is now supported, and it's LTR.

and MARK.md is updated to reflect both updates.