unify order of terms
terms in toc & in page had different order b/c the terms in toc were
sorted by the human name and the terms in the pages and in the summary
(tsv) were sorted by the ascii order, which currently only made a
difference in abbreviated names (2fa -> two-...).  now all follow
the ascii sorting order.