redirect nonexistent non-permalinks to their page
Now going on "ysmu/#ai" will have the same effect as the permalink
"ysmu/link/ai/" if JavaScript is enabled, ie, it'll redirect you
to the page that contains this term, to that part on the page (ie,
"ysmu/experimental/#ai" now). Unless it's already on the same page.

This works in the pages of all four stages, and redirects to any term
defined in any stage. So "ysmu/experimental/#a_candidate_term" works,
so does going vice versa; it's not limited to stable terms or the home
(stable) page.

It also supports short and long names; Both "ysmu/#ai" and
"ysmu/#artificial_intelligence" work.